requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
cryptography>=42.0.0
msal>=1.28,<2
flask-session>=0.8.0
requests>=2.31.0
